Move PFI Exemplar integration code into dedicated executable
