+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 members = [ 
-    "bindings/tbdex_uniffi",
+    "bindings/tbdex_uniffi", "crates/pfi_exemplar_integration_test",
     "crates/tbdex",
 ]
 default-members = [
