+++ crates/pfi_exemplar_integration_test/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "pfi_exemplar_integration_test"
+version = "0.1.0"
+edition = "2021"
+homepage.workspace = true
+repository.workspace = true
+license-file.workspace = true
+
+[dependencies]
+serde_json = { workspace = true }
+tbdex = { path = "../tbdex" }
+web5 = { workspace = true }
